# gateway/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "US Government Data API",
    "description": "Unified REST API providing access to 45 US government data endpoints across 13 categories including cybersecurity vulnerability intelligence. Real-time data from NVD, CISA KEV, EPSS, NHTSA, FDA, CFPB, FEMA, NOAA, USGS, SEC, BLS, USDA, EPA, HUD, FAA, CBP, SAM.gov, CourtListener, CDC, WHO, Federal Register, Grants.gov, OpenCorporates, and more.",
    "version": "1.1.0",
    "contact": {
      "name": "MARTC03",
      "url": "https://apify.com/martc03"
    }
  },
  "servers": [
    {
      "url": "https://govdata-api.netlify.app",
      "description": "Production (Netlify Proxy)"
    }
  ],
  "paths": {
...
    "/api/v1/disasters/fema-declarations": {
      "get": {
        "operationId": "searchFemaDeclarations",
        "summary": "Search FEMA Disaster Declarations",
        "description": "Search FEMA disaster declarations by state, type, and year range.",
        "tags": ["Disasters"],
        "parameters": [
          { "name": "states", "in": "query", "schema": { "type": "string" }, "description": "State codes (e.g., CA, TX)" },
          { "name": "incidentTypes", "in": "query", "schema": { "type": "string" }, "description": "Incident types" },
          { "name": "yearFrom", "in": "query", "schema": { "type": "integer" }, "description": "Start year" },
          { "name": "yearTo", "in": "query", "schema": { "type": "integer" }, "description": "End year" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/disasters/weather-alerts": {
      "get": {
        "operationId": "getWeatherAlerts",
        "summary": "Get NOAA Weather Alerts",
        "description": "Get real-time NOAA/NWS weather alerts by state, severity, and event type.",
        "tags": ["Disasters"],
        "parameters": [
          { "name": "states", "in": "query", "schema": { "type": "string" }, "description": "State codes" },
          { "name": "severity", "in": "query", "schema": { "type": "string" }, "description": "Alert severity" },
          { "name": "urgency", "in": "query", "schema": { "type": "string" }, "description": "Alert urgency" },
          { "name": "eventTypes", "in": "query", "schema": { "type": "string" }, "description": "Event types" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/disasters/earthquakes": {
      "get": {
        "operationId": "searchEarthquakes",
        "summary": "Search USGS Earthquakes",
        "description": "Search USGS earthquake data by magnitude, date range, and alert level.",
        "tags": ["Disasters"],
        "parameters": [
          { "name": "minMagnitude", "in": "query", "schema": { "type": "number" }, "description": "Minimum magnitude" },
          { "name": "maxMagnitude", "in": "query", "schema": { "type": "number" }, "description": "Maximum magnitude" },
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Start date" },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "End date" },
          { "name": "alertLevel", "in": "query", "schema": { "type": "string" }, "description": "Alert level" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
...
    "/api/v1/contracts/opportunities": {
      "get": {
        "operationId": "searchContractOpportunities",
        "summary": "Search SAM.gov Contract Opportunities",
        "description": "Search government contract opportunities (RFP/RFQ) on SAM.gov.",
        "tags": ["Contracts"],
        "parameters": [
          { "name": "keyword", "in": "query", "schema": { "type": "string" }, "description": "Search keyword" },
          { "name": "naicsCode", "in": "query", "schema": { "type": "string" }, "description": "NAICS code" },
          { "name": "postedFrom", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Posted from date" },
          { "name": "postedTo", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Posted to date" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/contracts/spending": {
      "get": {
        "operationId": "searchContractSpending",
        "summary": "Search USAspending Awards",
        "description": "Search government spending data from USAspending.gov.",
        "tags": ["Contracts"],
        "parameters": [
          { "name": "awardType", "in": "query", "schema": { "type": "string" }, "description": "Award type" },
          { "name": "agency", "in": "query", "schema": { "type": "string" }, "description": "Agency name" },
          { "name": "recipientName", "in": "query", "schema": { "type": "string" }, "description": "Recipient name" },
          { "name": "minAmount", "in": "query", "schema": { "type": "number" }, "description": "Minimum amount" },
          { "name": "maxAmount", "in": "query", "schema": { "type": "number" }, "description": "Maximum amount" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/contracts/entity-lookup": {
      "get": {
        "operationId": "lookupEntity",
        "summary": "Lookup SAM.gov Entity",
        "description": "Look up entity registration in SAM.gov by name.",
        "tags": ["Contracts"],
        "parameters": [
          { "name": "entityName", "in": "query", "schema": { "type": "string" }, "description": "Entity name" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/courts/opinions": {
      "get": {
        "operationId": "searchCourtOpinions",
        "summary": "Search Court Opinions",
        "description": "Search court case opinions from CourtListener.",
        "tags": ["Courts"],
        "parameters": [
          { "name": "query", "in": "query", "schema": { "type": "string" }, "description": "Search query" },
          { "name": "court", "in": "query", "schema": { "type": "string" }, "description": "Court identifier" },
          { "name": "dateAfter", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date after" },
          { "name": "dateBefore", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date before" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/courts/dockets": {
      "get": {
        "operationId": "searchCourtDockets",
        "summary": "Search Court Dockets",
        "description": "Search federal and state court dockets.",
        "tags": ["Courts"],
        "parameters": [
          { "name": "query", "in": "query", "schema": { "type": "string" }, "description": "Search query" },
          { "name": "court", "in": "query", "schema": { "type": "string" }, "description": "Court identifier" },
          { "name": "dateAfter", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date after" },
          { "name": "dateBefore", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date before" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/courts/judges": {
      "get": {
        "operationId": "searchJudges",
        "summary": "Search Federal Judges",
        "description": "Search the federal judge directory.",
        "tags": ["Courts"],
        "parameters": [
          { "name": "name", "in": "query", "schema": { "type": "string" }, "description": "Judge name" },
          { "name": "court", "in": "query", "schema": { "type": "string" }, "description": "Court" },
          { "name": "politicalAffiliation", "in": "query", "schema": { "type": "string" }, "description": "Political affiliation" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
...
  },
  "components": {
    "responses": {
      "SuccessResponse": {
        "description": "Successful response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": true },
                "data": { "type": "object" },
                "meta": {
                  "type": "object",
                  "properties": {
                    "tool": { "type": "string" },
                    "latencyMs": { "type": "integer" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string" },
                "meta": {
                  "type": "object",
                  "properties": {
                    "tool": { "type": "string" },
                    "latencyMs": { "type": "integer" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "RapidAPIKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Key"
      }
    }
  },
  "security": [
    { "RapidAPIKey": [] }
  ],
...
}